+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **Features**:
 
-* Support HTTP API v0.6.0, [PR-29](https://github.com/reduct-storage/reduct-js/pull/29)
+* Support Reduct Storage API v0.6, [PR-29](https://github.com/reduct-storage/reduct-js/pull/29)
 
 **Docs**: